Short_link/requirements.txt
alembic
fastapi[all]
psycopg2-binary
asyncpg
python-dotenv
sqlalchemy
sqlalchemy[asyncio]
uvicorn
redis
gunicorn
celery
pydantic
pyjwt
passlib